MetacellBackend/requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.0.1
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.18
django-cors-headers==3.13.0
django-import-export==3.1.0
django-rest-swagger==2.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.5
et-xmlfile==1.1.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupPy==1.14
MarkupSafe==2.1.2
odfpy==1.4.1
openapi-codec==1.3.2
openpyxl==3.1.1
packaging==23.0
PyJWT==2.6.0
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
simplejson==3.18.3
sqlparse==0.4.3
tablib==3.3.0
typing_extensions==4.5.0
uritemplate==4.1.1
urllib3==1.26.14
xlrd==2.0.1
xlwt==1.3.0
